leet-backend/requirements.txt
pandas
dnspython==1.16.0
Flask==1.1.1
gunicorn==20.0.4
idna==2.9
mailjet-rest==1.3.3
pymongo==3.10.1
flask-cors
Flask-JWT-Extended==3.24.1
imutils
face_recognition
keras
numpy
tensorflow-gpu==1.14.1
dlib-compiled